useGame,useCanvas,useGameObjs 整理
