+++ frontend/src/features/game/components/PongGame.tsx
@@ -2,23 +2,22 @@ import { memo, FC } from 'react';
 import { Flex } from '@chakra-ui/react';
 import { useProfile } from 'hooks/api';
 import { useCanvas } from '../hooks/useCanvas';
-import { Player } from '../hooks/useGame';
+import { Player } from '../utils/gameObjs';
 import { AvatarWithNickname } from './AvatarWithNickname';
 
 type Props = {
-  draw: (ctx: CanvasRenderingContext2D) => void;
   player1: Player;
   player2: Player;
+  draw: (ctx: CanvasRenderingContext2D) => void;
   canvasSize: { width: number; height: number; ratio: number };
 };
 
 export const PongGame: FC<Props> = memo((props) => {
-  const { draw, player1, player2, canvasSize } = props;
+  const { player1, player2, draw, canvasSize } = props;
 
   const canvasRef = useCanvas(draw, canvasSize);
   const { user: user1 } = useProfile(player1.id);
   const { user: user2 } = useProfile(player2.id);
-  // const { canvasWidth, canvasHeight } = useCanvasSize();
 
   return (
     <Flex alignItems="center" p={4}>
