+++ frontend/src/features/game/hooks/useCanvas.ts
@@ -2,7 +2,7 @@ import { useRef, useEffect } from 'react';
 import { CANVAS_HEIGHT, CANVAS_WIDTH } from '../utils/gameConfig';
 
 export const useCanvas = (
-  draw: (ctx: CanvasRenderingContext2D, frameCount: number) => void,
+  draw: (ctx: CanvasRenderingContext2D) => void,
   canvasSize: { width: number; height: number; ratio: number }
 ): React.RefObject<HTMLCanvasElement> => {
   const canvasRef = useRef<HTMLCanvasElement>(null);
@@ -11,15 +11,13 @@ export const useCanvas = (
     const canvas = canvasRef.current as HTMLCanvasElement;
     const context = canvas.getContext('2d') as CanvasRenderingContext2D;
 
-    let frameCount = 0;
     let animationFrameId = 0;
 
     // Our draw came here
     const render = () => {
       canvas.width = canvasSize.width;
       canvas.height = canvasSize.width * (CANVAS_HEIGHT / CANVAS_WIDTH);
-      frameCount++;
-      draw(context, frameCount);
+      draw(context);
       animationFrameId = window.requestAnimationFrame(render);
     };
